Trim upgrade targets and switch dorna_python, camera, vision to editable installs
Remove wizard, education, and dorna_lab setup scripts. Update upgrade list
in top-level setup.sh to dorna_python, vision, camera. Rewrite each remaining
sub-setup to be idempotent (git pull if $dir/.git exists, else fresh clone)
and use pip3 install -e . so future upgrades become git pull in place.
Vision clones the pro branch. README trimmed to match.

+++ dorna_python/setup.sh
@@ -6,20 +6,25 @@ dir="/home/dorna/Downloads/dorna_python"
 repo="https://github.com/dorna-robotics/dorna2-python.git"
 
 ########################
-#    clone the repo    #
+#    clone or pull     #
 ########################
-git clone $repo $dir
+if [ -d "$dir/.git" ]; then
+    cd $dir
+    git restore .
+    git pull
+else
+    rm -rf $dir
+    git clone $repo $dir
+    cd $dir
+fi
 
-# navigate to directory
-cd $dir
-git restore .
-git pull
-
-# install requirements
-pip3 install -r requirements.txt --break-system-packages
+# install requirements (if present)
+if [ -f requirements.txt ]; then
+    pip3 install -r requirements.txt --break-system-packages
+fi
 
 #################
 #    install    #
 #################
-# install package
-pip3 install . --upgrade --force-reinstall --break-system-packages
+# editable install
+pip3 install -e . --break-system-packages
