+++ vision/setup.sh
@@ -2,21 +2,32 @@
 ###################
 #    variables    #
 ###################
-current_dir="$( cd -- "$(dirname "$0")" >/dev/null 2>&1 ; pwd -P )"
 dir="/home/dorna/Downloads/vision"
 repo="https://github.com/dorna-robotics/dorna_vision"
+branch="pro"
 
 ########################
-#    clone the repo    #
+#    clone or pull     #
 ########################
-git clone $repo $dir
-# navigate to directory
-cd $dir
-git restore .
-git pull
+if [ -d "$dir/.git" ]; then
+    cd $dir
+    git restore .
+    git fetch origin
+    git checkout $branch
+    git pull
+else
+    rm -rf $dir
+    git clone -b $branch $repo $dir
+    cd $dir
+fi
 
-# install requirements
-pip3 install -r requirements.txt --break-system-packages
+# install requirements (if present)
+if [ -f requirements.txt ]; then
+    pip3 install -r requirements.txt --break-system-packages
+fi
 
-# install package
-pip3 install . --upgrade --force-reinstall --break-system-packages
+#################
+#    install    #
+#################
+# editable install
+pip3 install -e . --break-system-packages
